resolving ciclic dependency
