+++ .gitignore
@@ -4,3 +4,5 @@ database.db
 scielobooks.egg-info/
 dist/
 build/
+development.ini
+production.ini
