+++ scielobooks/security.py
@@ -1,10 +1,10 @@
 
-from models import users_models as users 
+from users import models as user_models
 
 def groupfinder(userid, request):
     
-    USERS = request.rel_db_session.query(users.User.id).all()
+    USERS = request.rel_db_session.query(user_models.User.id).all()
     if userid in [user[0] for user in USERS]:
-        group_name = request.rel_db_session.query(users.User).get(userid).group.name
+        group_name = request.rel_db_session.query(user_models.User).get(userid).group.name
         
         return [group_name]
